Created Contribution Guidelines (read commit desc)
For those with a new fork trying to pull request back into master: This should not interfere with any asset-adding pull requests that others want to do. If you're trying to do a pull request and there's this file showing up as a conflict, just continue with the pull request and things should work out fine. Worst comes to worst is that you'll have to merge my fork back into yours first, and re-upload the files you added.